Move some container related methods and structs to smaller files
